Analysis functions

MolSimToolkit.distances — Method

distances(simulation, indexes1, indexes2)

Function that calculates the distance between the centers of mass of two selections in a simulation.

The selections are defined by the indexes1 and indexes2 vectors, which are the indexes of the atoms.

Example

julia> import PDBTools

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> simulation = Simulation(Testing.namd_pdb, Testing.namd_traj);

julia> i1 = PDBTools.selindex(atoms(simulation), "protein and residue 1");

julia> i2 = PDBTools.selindex(atoms(simulation), "protein and residue 15");

julia> distances(simulation, i1, i2)
5-element Vector{Float64}:
 23.433267858947584
 30.13791365033211
 28.48617683945202
 27.92740141686934
 23.235012287435566

MolSimToolkit.center_of_mass — Method

center_of_mass(
    indexes::AbstractVector{Int};
    simulation::Simulation,
    positions::FramePositions,
    iref::Int = max(1, div(length(indexes),2)),
)

Calculate the center of mass of a selection of atoms in a simulation given the positions. The selection is defined by the indexes vector, which is the indexes of the atoms.

The iref parameter is the index of the reference atom. The center of mass is calculated by first computing the minimum-image of all atoms relative to this atom. By default, it is the atom closest to the middle of the indexes vector.

julia> import PDBTools 

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> simulation = Simulation(Testing.namd_pdb, Testing.namd_traj);

julia> protein_indexes = PDBTools.selindex(atoms(simulation), "protein");

julia> coor = positions(current_frame(simulation));

julia> cm = center_of_mass(protein_indexes, simulation, coor)
3-element Point3D{Float64} with indices SOneTo(3):
 -3.7290442807974906
 -1.5339226637687564
  1.960640754560446