Developer zone

Trajectory

MolSimToolkit.Trajectory — Type

Trajectory(file::String; first=1, last=nothing, step=1)

Creates a new Trajectory object from a file. If first, last, and step are not specified, the Trajectory will iterate over all frames in the file.

A Trajectory object represents a trajectory file. It can be iterated over to obtain the frames in the trajectory. The Trajectory object is a mutable struct that contains the following data, that can be retrived by the corresponding functions:

frame_range(::Trajectory): the range of frames to be iterated over

frame_index(::Trajectory): the index of the current frame in the trajectory

frame(::Trajectory): the current frame in the trajectory

length(::Trajectory): the number of frames to be iterated over in the trajectory file, considering the current range

raw_length(::Trajectory): the number of frames in the trajectory file

The Trajectory object can also be manipulated by the following functions:

close(::Trajectory): closes the trajectory file

restart!(::Trajectory): restarts the iteration over the trajectory file

currentframe(::Trajectory): returns the current frame in the trajectory

nextframe!(::Trajectory): reads the next frame in the trajectory file and returns it. Moves the current frame to the next one.

set_frame_range!(::Trajectory; first, last, step): resets the range of frames to be iterated over.

One important feature of the Trajectory object is that i can be iterated over, frame by frame. The pairs iterator can also be used to iterate over the frames, returning a tuple with the frame index and the frame itself. The enumerate iterator can also be used to iterate over the frames, returning a tuple with the frame counter and the frame itself.

Examples

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> trajectory = Trajectory(Testing.namd_traj, first = 2, step = 2, last = 4);

julia> for frame in trajectory 
           @show frame_index(trajectory)
           # show x coordinate of first atom 
           @show Positions(frame)[1].x
       end
frame_index(trajectory) = 2
((Positions(frame))[1]).x = 5.912472724914551
frame_index(trajectory) = 4
((Positions(frame))[1]).x = 7.346549034118652

julia> for (i, frame) in pairs(trajectory)
           @show i, frame_index(trajectory)
       end
(i, frame_index(trajectory)) = (2, 2)
(i, frame_index(trajectory)) = (4, 4)  

julia> for (i, frame) in enumerate(trajectory)
           @show i, frame_index(trajectory)
       end
(i, frame_index(trajectory)) = (1, 2)
(i, frame_index(trajectory)) = (2, 4)

Base.close — Method

close(trajectory::Trajectory)

Closes the trajectory file.

Base.length — Method

length(trajectory::Trajectory)

Returns the number of frames to be iterated over in the trajectory file, considering the current frame range.

MolSimToolkit.currentframe — Method

currentframe(trajectory::Trajectory)

Returns the current frame in the trajectory.

MolSimToolkit.frame — Method

frame(trajectory::Trajectory)

Returns the current frame in the trajectory.

MolSimToolkit.frame_index — Method

frame_index(trajectory::Trajectory)

Returns the index of the current frame in the trajectory.

MolSimToolkit.frame_range — Method

frame_range(trajectory::Trajectory)

Returns the range of frames to be iterated over.

MolSimToolkit.nextframe! — Method

nextframe!(trajectory::Trajectory)

Reads the next frame in the trajectory file and returns it. Moves the current frame to the next one in the range to be considered (given by frame_range(trajectory)).

MolSimToolkit.path — Method

path(trajectory::Trajectory)

Returns the path to the trajectory file.

MolSimToolkit.raw_length — Method

raw_length(trajectory::Trajectory)

Returns the number of frames in the trajectory file.

MolSimToolkit.restart! — Method

restart!(trajectory::Trajectory)

Restarts the iteration over the trajectory file.

MolSimToolkit.set_frame_range! — Method

set_frame_range!(trajectory::Trajectory; first=1, last=nothing, step=1)

Resets the frame range to be iterated over. This function will restart the iteration from the first frame of the new range.

Positions

MolSimToolkit.Point3D — Type

Point3D{T}

A point in 3D space with coordinates x, y, and z of type T.

MolSimToolkit.Positions — Type

Positions{T<:AbstractArray}

Container for the positions of a set of atoms. The positions are stored in a matrix, where each column corresponds to the coordinates of an atom. The container is used such that using the coodinates from a Chemfiles.Frame is transparent to the user, and the coordinates can be accessed as positions[i] where i is the index of the atom.

The coordinates of the atom can be accessed as positions[i].x, positions[i].y, and positions[i].z.

Example

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> trajectory = Trajectory(Testing.namd_traj);

julia> frame = currentframe(trajectory);

julia> positions = Positions(frame);

julia> positions[1]
3-element Point3D{Float64} with indices SOneTo(3):
  5.912472724914551
 10.768872261047363
 28.277008056640625

julia> positions[1].x
5.912472724914551

julia> positions[1].y
10.768872261047363

julia> positions[1].z
28.277008056640625

Unit cell

MolSimToolkit.UnitCell — Type

UnitCell{T}

Container for the unit cell matrix. The container is used such that using the unit cell from a Chemfiles.Frame is transparent to the user, without any further conversions.

Example

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> trajectory = Trajectory(Testing.namd_traj);

julia> frame = currentframe(trajectory);

julia> unitcell = UnitCell(frame)
UnitCell{Float64}
  47.411   0.000   0.000
   0.000  47.411   0.000
   0.000   0.000  87.798

julia> positions = Positions(frame);

julia> wrap(positions[1], positions[2], unitcell)
3-element Point3D{Float64} with indices SOneTo(3):
  5.912472724914552
 10.768872261047363
 28.27700805664062

Wrap coordinates

MolSimToolkit.wrap — Function

wrap(x::Point3D, xref::Point3D, unitcell::UnitCell) 
wrap(x, xref, unit_cell_matrix::SMatrix{N,N,T}) where {N,T}
wrap(x, xref, sides::AbstractVector)

Wraps the coordinates of point x such that it is the minimum image relative to xref. The unit cell may be given as a UnitCell structure, a static matrix of size (N,N) or as a vector of length N.

MolSimToolkit.wrap_to_first — Function

wrap_to_first(x::Point3D, unitcell::UnitCell)
wrap_to_first(x, unit_cell_matrix)

Wraps the coordinates of point x such that the returning coordinates are in the first unit cell with all-positive coordinates. The unit cell may be given as a UnitCell structure or as a static matrix of size (N,N).

Example

julia> using MolSimToolkit, StaticArrays

julia> unitcell = UnitCell(@SMatrix[10 0 0; 0 10 0; 0 0 10])
UnitCell{Int64}
  10.000   0.000   0.000
   0.000  10.000   0.000
   0.000   0.000  10.000

julia> wrap_to_first(Point3D(15, 13, 2), unitcell)
3-element SVector{3, Float64} with indices SOneTo(3):
 5.0
 3.0000000000000004
 2.0