Analysis functions

MolSimToolkit.distances — Method

distances(trajectory, indexes1, indexes2; masses=nothing)

Function that calculates the distance between the center of masses of the selections in a trajectory, or the center of coordinates if masses are not provided.

julia> using PDBTools

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> trajectory = Trajectory(Testing.namd_traj);

julia> atoms = readPDB(Testing.namd_pdb);

julia> masses = mass.(atoms);

julia> i1 = selindex(atoms, "protein and residue 1");

julia> i2 = selindex(atoms, "protein and residue 15");

julia> distances(trajectory, i1, i2; masses = masses)
5-element Vector{Float64}:
 23.433267858947584
 30.13791365033211
 28.48617683945202
 27.92740141686934
 23.235012287435566

MolSimToolkit.center_of_mass — Method

center_of_mass(
    positions::Positions, indexes::AbstractVector{Int}; 
    masses::Union{Nothing,AbstractVector{<:Real}} = nothing, 
    unitcell::Union{Nothing,UnitCell} = nothing,
    iref::Int = indexes[1],
)

Calculate the center of coordinates of a selection of atoms, or optionally the center of mass if masses are provided.

If a unitcell is provided, the positions are wrapped relative, by default, to the first atom in the selection. This can be changed by setting iref to the index of the atom to be used as the reference.

julia> using MolSimToolkit, MolSimToolkit.Testing

julia> using PDBTools

julia> atoms = readPDB(Testing.namd_pdb);

julia> protein_indexes = selindex(atoms, "protein");

julia> protein_masses = mass.(atoms);

julia> trajectory = Trajectory(Testing.namd_traj);

julia> positions = Positions(currentframe(trajectory));

julia> unitcell = UnitCell(currentframe(trajectory));

julia> cm = center_of_mass(positions, protein_indexes, masses=protein_masses, unitcell=unitcell)
3-element Point3D{Float64} with indices SOneTo(3):
 -3.7290442807974906
 -1.0539450244972206
 24.513632812542127