System setup

Packmol Input Creator

This module helps the setup of a Packmol input file, by computing the number of molecules and sizes necessary to build different systems. Currently (as of version 1.2.0) the module supports the construction of systems of a solute solvated by a mixture of two solvents.

How to use it

julia> using MolSimToolkit.PackmolInputCreator

Setting up the system properties

Here, we setup a system of a polymer solvated by water and ethanol. The densities as a function of the molar fraction of ethanol are available in a data table:

density_table = [
# x ethanol     density
   0.0000       0.9981
   0.0416       0.9820
   0.0890       0.9685
   0.1434       0.9537
   0.2066       0.9369
   0.2809       0.9151
   0.3695       0.8923
   0.4769       0.8685
   0.6098       0.8450
   0.7786       0.8195
   1.0000       0.7906
]

Next, we initialize the system data structure, given the PDB files of one molecule of the polymer (poly_h.pdb), and one molecule of water and ethanol:

# Directory of test files
test_dir = MolSimToolkit.PackmolInputCreator.PackmolInputCreatorDirectory*"/test"
# Construction of system data structure
system = SolutionBoxUSC(
    solute_pdbfile = "$test_dir/data/poly_h.pdb",
    solvent_pdbfile = "$test_dir/data/water.pdb",
    cossolvent_pdbfile = "$test_dir/data/ethanol.pdb",
    density_table = density_table,
)

Here, SolutionBoxUSC stands for Solute (U), Solvent (S), and Cossolvent (C).

Finally, we can generate an input file for Packmol with:

write_packmol_input(
    system; 
    concentration = 0.5, cunit = "x", 
    margin = 20.0, 
    input = "box.inp",
    output = "system.pdb"
)

The concentration can be given in molar fraction (x), molarity (mol/L), or volume fraction (vv).

The input parameter provides the name of the input file for Packmol that will be generated.

The margin parameter sets the size of the box, which will take into consideration the maximum and minimum dimensions of the solute. Alternatively, the size of the box can be provided explicitly with the box_sides = [ a, b, c ] parameters, where a, b, and c are the lengths of the box in each dimension.

Running Packmol

Packmol can be run directly from within Julia using the Packmol Julia package:

using Packmol
run_packmol("./box.inp")

If everything runs correctly, the output file system.pdb will be generated.

Help entries

MolSimToolkit.PackmolInputCreator.convert_concentration — Method

convert_concentration(
    system::SolutionBoxUSC,
    input_concentration, 
    units;
    density, # of the solution, required if the input concentration is mol/L
)

Convert concentration from one unit to another. The input concentration is given in input_concentration, and the unit conversion is given by units keyword, that can be one of the following pairs:

"mol/L" => "x": from molarity to molar fraction

"mol/L" => "vv": from molarity to volume fraction

"x" => "mol/L": from molar fraction to molarity

"x" => "vv": from molar fraction to volume fraction

"vv" => "mol/L": from volume fraction to molarity

"vv" => "x": from volume fraction to molar fraction

MolSimToolkit.PackmolInputCreator.write_packmol_input — Method

write_packmol_input(
    system::SolutionBoxUSC;
    concentration::Real, 
    cunit="mol/L",
    input="box.inp",
    output="system.pdb",
    # box size
    box_sides::AbstractVector{<:Real}, # or
    margin::Real
)

Function that generates an input file for Packmol. By default, the concentrations is given in mol/L, but it can also be given in molar fraction "x" or volume fraction "vv", using cunit="x" or cunit="vv".

The box sides are given in Ångströms, and can be provided as a vector of 3 elements. Alternativelly, the margin can be provided, and the box sides will be calculated as the maximum and minimum coordinates of the solute plus the margin in all 3 dimensions.

MolSimToolkit.PackmolInputCreator.SolutionBoxUSC — Type

SolutionBoxUSC(; 
    solute_pdbfile::String, 
    solvent_pdbfile::String,
    cossolvent_pdbfile::String,
    density_table::Matrix{Float64},
)

Setup a system composed of a solute (U) a solvent (S) and a cossolvent (C).